+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/
 [semantic versioning]: https://semver.org/spec/v2.0.0.html
 
-## [Unreleased]
+## [0.6.2] - 2022-08-24
 
 ### Added
 
@@ -111,6 +111,8 @@ The format is based on [Keep a Changelog], and this project adheres to
 [0.5.0]: https://github.com/dogmatiq/imbue/releases/tag/v0.5.0
 [0.5.1]: https://github.com/dogmatiq/imbue/releases/tag/v0.5.1
 [0.6.0]: https://github.com/dogmatiq/imbue/releases/tag/v0.6.0
+[0.6.1]: https://github.com/dogmatiq/imbue/releases/tag/v0.6.1
+[0.6.2]: https://github.com/dogmatiq/imbue/releases/tag/v0.6.2
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD